convometrics-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
anthropic
httpx
apscheduler
python-multipart
email-validator
